Move diffusion to examples
